requirements.txt
numpy
pillow
docopt
tornado==4.5.3
requests
h5py
python-socketio
flask
eventlet
moviepy
pandas

tensorflow==1.8.0


#picamera
Adafruit_PCA9685